requirements.txt
flask
pandas
python-dotenv
sqlite3
openpyxl
